Rename common data storage (matches naming in the runtime)
